+++ google/cloud/bigtable/admin_client.h
@@ -308,37 +308,6 @@ std::shared_ptr<AdminClient> CreateDefaultAdminClient(std::string project,
 
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace bigtable
-namespace bigtable_internal {
-GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
-
-/**
- * A helper class for `DefaultAdminClient` to process `Options` before they are
- * used to initialize its `BigtableTableAdminConnection`. The point is to save
- * the creation of a background thread. The class is factored out for testing
- * purposes.
- *
- * `TableAdmin::WaitForConsistency()` requires a CQ to run the
- * polling loop. We would like to run this CQ on the same background threads
- * that the Connection uses, instead of spinning off an extra thread, just
- * for this purpose.
- *
- * If the user supplies their own threads, we can use those to run the
- * polling loop.
- *
- * Otherwise, we will create and store the background threads
- * in this class. Then we will tell the Connection to use our threads for
- * its background work.
- */
-struct AdminClientParams {
-  explicit AdminClientParams(Options opts);
-
-  CompletionQueue cq;
-  std::unique_ptr<BackgroundThreads> background_threads;
-  Options options;
-};
-
-GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
-}  // namespace bigtable_internal
 }  // namespace cloud
 }  // namespace google
