+++ google/cloud/bigtable/google_cloud_cpp_bigtable.bzl
@@ -61,6 +61,7 @@ google_cloud_cpp_bigtable_hdrs = [
     "instance_config.h",
     "instance_list_responses.h",
     "instance_update_config.h",
+    "internal/admin_client_params.h",
     "internal/async_bulk_apply.h",
     "internal/async_longrunning_op.h",
     "internal/async_poll_op.h",
@@ -138,6 +139,7 @@ google_cloud_cpp_bigtable_srcs = [
     "instance_admin_client.cc",
     "instance_config.cc",
     "instance_update_config.cc",
+    "internal/admin_client_params.cc",
     "internal/async_bulk_apply.cc",
     "internal/async_row_sampler.cc",
     "internal/bulk_mutator.cc",
